--- Not necessary to consider for release 0.3.3. ---
2002-01-02  Alexander R Angas  <wgd@internode.on.net>

	* src/main.c:
	* src/engine/kgraphics.c:
	* src/engine/klists.c:
	* src/engine/kmovement.c:
	* src/engine/kscripts.c:
	* src/engine/savegame.c:
	* src/engine/savegame.cfsml:
	* src/engine/scriptdebug.c
	* src/engine/vm.c:
	* src/gfx/menubar.c:
	* src/gfx/widgets.c:
	* src/gfx/resource/sci_pic_0.c:
	* src/include/soundserver.h:
	* src/scicore/resource.c:
	* src/sound/oldmidi.c:
	* src/include/win32/usleep.h: Warnings fixes.

--- Not necessary to consider for release 0.3.3. ---

+++ src/engine/klists.c
@@ -74,7 +74,7 @@ sane_listp(state_t *s, heap_ptr address)
   if (last==-1)
     return 0;
 
-  return sane_nodep(s, seeker);
+  return sane_nodep(s, (unsigned short)seeker);
 }
 
 void
@@ -369,7 +369,7 @@ kNodeValue(state_t *s, int funct_nr, int argc, heap_ptr argp)
 
 	if (!sane_nodep(s, UPARAM(0)))
 	  SCIkwarn(SCIkERROR,"List node at %04x is not sane anymore!\n", PARAM(0));
-  
+
 	a = UPARAM(0) + LIST_NODE_VALUE;
 
 	s->acc=UGET_HEAP(a);
